Fix assert_await* methods of an autospec coroutine
When calling any assert_awaited() method on a coroutine mock returned by
create_autospec(), the method actually called was assert_not_awaited().

This is because we were using a closure in a for-loop to retrieve the
mehod of the wrapped coroutine, and thus the wrapper was always using
the last value of the variable of the loop...

I added a test to ensure it doesn't happen again.