Bind patches to the coroutine instance
This fixes #121. The problem is that if a patched coroutine runs
multiple times concurrently, they used to share the patch object, which
affected the concurrent runs.

It had the side effect of pinning the patch mock to the coroutine
function, hence each call received the same mock.

TODO document the behavior and the quirks.